[level 0] Title: 제곱수 판별하기, Time: 0.01 ms, Memory: 4.15 MB -BaekjoonHub

+++ 프로그래머스/0/120909. 제곱수 판별하기/제곱수 판별하기.cpp
@@ -5,6 +5,6 @@
 using namespace std;
 
 int solution(int n) {
-    int s = (int)sqrt(n);
+    int s = sqrt(n);
     return s*s==n ? 1 : 2;
 }
